Initial check-in
This is a personal accounting program that I wrote in the 1978 for UNIX,
translated to DOS a few years later (so I could run it at home) and
finally rewrote in Java (in 2010) as practice for an intro-to-Java class
I was asked to teach.

It has budgets, records expenses and deposits, allows drill-downs,
displays summaries, and keeps all of its data in an easily edited ASCII
text file.  I don't claim it is great, but it has met my needs for a
very long time.